Bugfix: Fix regression from #5739 for passing plugin name and reader plus add test (#6013)
# Fixes/Closes

Issue with 0.4.18rc2 reported on zulip by @adamltyson 

https://napari.zulipchat.com/#narrow/stream/212875-general/topic/reader.20plugin.20issue.20with.200.2E4.2E18

Closes #6012

# Description

In merged commit:
https://github.com/napari/napari/commit/6847eeef0d036b213c896577508d83aae3e40146
the method of checking whether a plugin is npe2 returns None if a reader
is passed in addition to the plugin name, e.g. `my-plugin.some_reader`
vs. `my-plugin`
This is a regression for the case of plugins with multiple readers, as
reported on zulip:

https://napari.zulipchat.com/#narrow/stream/212875-general/topic/reader.20plugin.20issue.20with.200.2E4.2E18

This PR just changes the logic of the comparison so use `startswith` so
just the plugin name is compared, rather than the entire string. This
fixes the regression.
Also, I add a test for this condition that fails on main, but passes
with this fix.

# References
See:
https://napari.zulipchat.com/#narrow/stream/212875-general/topic/reader.20plugin.20issue.20with.200.2E4.2E18