Feature: polygon drawing in the Labels layer (#5806)
# Description

This PR implements a new way of drawing in the Labels layer by creating
polygons instead of just using the brush. It is quite useful when
working with objects that have straight lines or in other scenarios when
it is easier to specify an object mask by following its perimeter.

It is implemented by introducing a new mode in the Labels layer called
`DRAW_POLYGON`:

![image](https://user-images.githubusercontent.com/5459767/236327502-709964be-47aa-45ae-9560-dc7226725aad.png)

Polygon drawing is controlled by the following actions:

- Create a new point (left mouse click)
- Delete the latest point of the polygon (right mouse click)
- Complete the polygon creation and adding the polygon mask to the
labels (`Enter`, which can be specified in the Labels shortcuts or clicking within
20 screen pixels of the first vertex.)
- Cancel the polygon creation (`Escape`, can be specified in the Labels
shortcuts)

It is important to note that switching modes does not complete or cancel
the polygon creation, which is done on purpose to be able to use
`PAN_ZOOM` during the process (it is extremely useful when working with
large objects as in the video below).

The color of the polygon matches the current label color and it takes
opacity from the layers opacity. There is only one exception: when you
use background color, the polygon is always transparent and has red
boundaries.

Also, the polygon implementation is completely independent of the Shapes
layer and uses its own polygon visualization based on an overlay, which
has better performance and, more importantly, is substantially easier
and more straightforward to implement.

https://user-images.githubusercontent.com/5459767/236333440-9bf07491-c328-4509-91ef-3ae07fe7950c.mp4

https://user-images.githubusercontent.com/5459767/236326736-902322b1-bde5-46be-b8c2-d5c2e0ea42bd.mp4


## Type of change
- [X] New feature (non-breaking change which adds functionality)

# How has this been tested?
- [X] all tests pass with my change

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Wouter-Michiel Vierdag <w-mv@hotmail.com>