Show partial shortcut while editing.
And automatically accept/validate shortcuts when valid.

 - When receiving a modifier only event, use the event Modifiers to show
   the current partial shortcuts.

 - we need to handle keyrelease events otherwise the following sequence
   is incorrect.

```
Action          –  Shortcut shown
---------------------------------
Press Ctrl      –  Ctrl
Also Press Alt  –  Ctrl-Alt
Release Ctrl    –  Ctrl-Alt
```

Though now that we handle release, the user cant press a shortcut and
then validate with enter as as soon as the user release the shortcut
the QLineEdit will become empty.

Thus validate as soon a shortcut is valid, and contain a non-modifier
key.

This in turn make it impossible to assign a modifiers only keybinding,
I assume this is quite rare, but we can use on of the keys that are
invalid as shortcut to validate modifiers only shortcuts. I choose
Qt.Return, aka `Enter`.

I'm not quite confident on the Return/Enter, and the way to assign
modifiers only shortcuts.

There is also this whole Meta/Alt/Ctrl are not the same on MacOS.
I did not extract the logic into naprai/utils/interactions as those are
Qt Specific, moreover the values of QtEvent.Modifiers are not the same
than Qt.Keys.

You will note that there is one weirdness in the order of modifiers when
typing these, but I believe this will be fixed upstream.

Closes #6047