Async refresh when experimental setting is on (#6008)
# Description

This makes `refresh` attempt async slicing by default when its
corresponding experimental flag is on. In that case, the many internal
calls to `Layer.refresh` will be executed asynchronously via the viewer
rather than directly in the layer. This is especially helpful when a
layer is initialized and added to the viewer, which can trigger many
blocking calls to `refresh` some of which can now be cancelled. As
`Layer._slice_dims` is always on the sync path, that forces the refresh
to be synchronous.

In most usage of napari this should be enough to avoid #5591 because if
the experimental flag is on and the layer supports async, then all
slicing should occur on the slicing thread. If the flag is off or the
layer does not support async, then all slicing occurs on the main
thread.

However, #5591 could still be triggered if a user flips the experimental
flag while async slicing is occurring. So this PR does not entirely
close #5591, though something like #5610 should.

# References

Related to #5591 but does not close it.