Use a shader for low discrepancy label conversion (#3308)
The pull request will move the current label color calculation onto the
GPU.

This is the result of dozens of hours of hacking between @perlman,
@jni, and @brisvag. It moves the mapping from label space to 0-1
(which can then be colormapped) to the shader.

It also adds hashing in OpenGL for direct color mode, which avoids
many of the collisions we were seeing in that mode when mapping
labels to 0-1 and then using a custom colormap.

Closes #204 (count those digits!!!!)
Closes #4384 (which you'll see has been hit by many people  it's a big one)
Closes #4864
Closes #4942
Closes #6077
Closes #6010

This converts labels to np.float32, so labels above `2**23` can be aliased
together. But this limitation seems preferable to existing issues.

- [X] Implement shader for 3D label layers (ISO categorical)
- [X] Implement shader for 2D label layers
- [X] Fix the CPU-based color computation (for paint color preview &
layer preview)
- [X] 3D iso categorical view
- [x] Test function of directly mapped colors

- [ ] Address uint64 labels. Labels are currently converted to float32
- [ ] Factor `seed` into a `uniform`

- [X] This change requires a documentation update

- [X] Additional tests added to confirm consistent coloring between GPU
& CPU
- [X] Manual testing of the various examples (`examples/**labels*.py`)

---------

Co-authored-by: Juan Nunez-Iglesias <juan.nunez-iglesias@monash.edu>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>