Return full colormap dict in get_state (#5782)
# Description
When working with custom colormaps, reader and writers trying to use
that colormap will fail because the only information returned by
`Layer.get_state` is the colormap name. This works fine with builtin
colormaps, but anything custom will fail due to `[unnamed colormap 0]`
not existing.

By returning the full colormap model (in dict form) we can also account
for this.

Plugin writers and readers using `colormap` will likely break (like our
napari-svg) because they assume it to be a string.

Note that this includes #5805. which is a lot smaller and just a bug
fix.

# References
<!-- What resources, documentation, and guides were used in the creation
of this PR? -->

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

---------

Co-authored-by: Kyle I S Harrington <kephale@users.noreply.github.com>