Fix point selection highlight (#5737)
# Description
Fix some tests for point selection, and fix highlight not being updated
correctly when changing zoom.


To reproduce the issue on main:

```py
import napari
import numpy as np
v = napari.Viewer()
pts = np.array([[0, 0], [10, 10]])
pl = v.add_points(pts, scale=[5, 5])
pl.mode = 'select'
```

Then select points (manually or with a box) and zoom in and out, and
repeat. You'll see that unless you hover a point, the hihglight
thickness is not properly updated with the zoom. This PR ensures that
zooming always triggers an update to the highlight thickness.

---------

Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>