Fix `_add_sample` sample data callback to show all readers (#6058)
While working on #4865, specifically converting
[`test_sample_data_triggers_reader_dialog`](https://github.com/napari/napari/blob/007712f0647d4f94cda02b79e58a08273846ac94/napari/_qt/menus/_tests/test_file_menu.py#L12)
to use the new app-model, I noticed only one plugin was listed in the pop
up (see second to last bullet point here:
https://github.com/napari/napari/pull/4865#issuecomment-1285035203).
I checked and this was also the case in main, and tracked the issue to
`handle_gui_reader`.

When opening a sample data URI from the GUI, if there is more than one
reader, we want to use the `handle_gui_reader`, so the reader choice pop
up comes.

Currently the `_add_sample` callback that opens sample data from the
menu will remove the plugin that provided the sample data, from possible
readers.

The plugin that provided the sample data was passed via `plugin_name`
parameter to `handle_gui_reading`, implying that we already tried to
open the sample with this plugin. However, `handle_gui_reading` is only
called if multiple readers are available (which means no readers were
tried), so we do not want to remove this plugin from possible readers.