Extend "No Qt bindings found" error message with details about conda (#6095)
# Fixes/Closes

Extend "No Qt bindings found" error message with details about conda.

This is porting changes we had to add (due to urgency) directly in the
conda-forge feedstock (see
https://github.com/conda-forge/napari-feedstock/pull/52).

# Description

We have changed how we package `napari` for conda, and it doesn't depend
on PyQt or PySide anymore. Instead, the user needs to specify which one
they want (see
https://github.com/conda-forge/napari-feedstock/issues/47). This has
been addressed in the documentation, but there might be people still
relying on the "it just works" behavior of `conda install napari`.

For those, we will have a nicer error message that will tell them how to
fix the absence of Qt bindings. In 0.4.18 this didn't work on macOS due
to the `qtpy` import in `napari.__main__._maybe_rerun_with_macos_fixes`,
but this is fixed in `main`. This is why the patch in the 0.4.18
feedstock is different. In this PR we only need to extend the error
message already present in `napari._qt.__init__`.

# References

- https://github.com/conda-forge/napari-feedstock/pull/52
- https://github.com/conda-forge/napari-feedstock/issues/47

## Type of change
- [X] Enhancement (non-breaking improvements of an existing feature)

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [X] Created a conda installation for napari, force removed pyqt and
tried to run `napari`. The following message shows up:

```
ImportError: No Qt bindings could be found.

napari requires either PyQt5 (default) or PySide2 to be installed in the environment.

With pip, you can install either with:
  $ pip install -U 'napari[pyqt]'
  $ pip install -U 'napari[pyside]'

With conda, you can install either with:
  $ conda install pyqt
  $ conda install pyside2
```

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>