Update PR editing script to remove html.
This was more complicated than I anticipated, in particular because
GitHub documentation seem to be wrong in a number of places.

Action environment also seem to differ for intra- and inter- repository
PRs as well as org- vs user- base repo.

This successfully stripped html from this PR:

    https://github.com/Carreau/napari/pull/4

Though it was in intra-repo with user-based repo, so I'm not sure how this will behave
in an inter-repo where the base-repo is an org.