Install napari from repository in docker image (#6097)
# Fixes/Closes

Closes https://github.com/napari/napari/issues/6096

# Description

We have a workflow that runs on each push and publishes a docker image
to ghcr.io. This includes tagged releases.

However, the Dockerfile installs napari from PyPI, which will always be
lagging. Hence, the docker container for 0.4.18 ships 0.4.17... and so
do all the previous commits since 0.4.17 excluded 😂 We have been
publishing the same Docker image for each push for a while and maaaybe
we should delete a few (although layers are cached, so not so bad).

# References

- `build-args`:
https://github.com/docker/build-push-action/tree/v4/#inputs
- ARG vs ENV:
https://vsupalov.com/docker-arg-env-variable-guide/#arg-and-env-availability

## Type of change

- [x] Bug-fix (non-breaking change which fixes an issue)

# How has this been tested?

- [x] CI passes and reports correct version  

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have added tests that prove my fix is effective or that my
feature works