Add basic tests for the `ScreenshotDialog` widget and fixes (#6057)
# Fixes/Closes

closes #6078 

# Description

Adds some tests to the `ScreenshotDialog` as a continuation of the
effort started at #5864 to increase Qt widgets code coverage and fixes
some errors with the non-native version (the one used when launching
napari from and IPython console)

## Notes and questions

* The tests call the dialog `show` method and also sets the non-native
option so the tests here only cover the case when napari is being
launched from an IPython console. The testing of the OS native version
of the dialog is more involved.
* Checking the code related with the dialogs available I noticed that
some of them (files/widgets classes) have the `qt_*` /`Qt` prefix or the
`_dialog`/`Dialog` suffix but others don't (for example
`screenshot_dialog.py/ScreenshotDialog` vs `qt_about/QtAbout`). Is that
something worthy to standardize or maybe there is already some logic
behind the naming?

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>