Fix oblique button by chekcing if action is generator (#6145)
This Pr fixes a bug that was claimed to be tested in #4265, but it does
not happens (action is dataclass so never will be generator).

As having `orient_plane_normal_along_view_direction` as the generator is
important then I add a separate function
`orient_plane_normal_along_view_direction_no_gen` for this.

- [x] Bug-fix (non-breaking change which fixes an issue)

- [x] Manually

Closes #6040