ENH Adds `LayerList` context (#6165)
Adds `LayerList` context. Currently we only have a `Selection[Layer]`
context. For the file menu (#4865) we needed `LayerList` because we
needed the number of layers present (NOT number of layers selected) and
could not obtain this from `Selection[Layer]`. The number of layers
present is required for the `enablement` status of the
`CommandId.DLG_SAVE_LAYERS` command.

Note this was originally part of #4865 but move to separate PR to aid
review and git blame.

# References

ref: https://github.com/napari/napari/pull/4865#discussion_r1292895534