Ensure visual is updated when painting into zarr array (#6112)
Fixes #6079

It turns out that the caching behaviour introduced in #5732 depends on the
slice data being updated by painting. This works out for NumPy arrays because
the slice data is a view of the original data, so updating the original (as
painting does) updates the slice. However, when the data is a zarr or
tensorstore array, the slice is a NumPy copy of the original data, so the
caching mechanism believes that nothing has changed and the display is not
updated.

This adds tests for the behaviour and fixes it by painting directly into the
slice data if the data array is not a NumPy array. It's a bit of a bandaid fix
but it works and is
[endorsed](https://github.com/napari/napari/issues/6079#issuecomment-1648781280)
by our slicing expert @andy-sweet. :joy:

(I've also made a couple of drive-by updates to the code because some
methods are no longer used in the code after #5732 but that was missed at the
time.)

## Type of change

- [x] Bug-fix (non-breaking change which fixes an issue)

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>