Layer data events before and after (#6178)
Closes #6177

This PR fixes the issue of `layer.events.data` not emitting in a
consistent manner. For example in case of adding a shape, an emit would
happen at the start of drawing a shape, while for removing data it would
happen after data was removed. This PR fixes the issue by changing
`ActionType` in present participle and past participle and emitting
before and after `layer.data` has been edited.
Furthermore, the way `layer.events.data` emits in the `data.setter` of
`Shapes` and `Points` has changed. The behaviour is as follows:

1. If `layer.data` is populated and new data is not empty use
`ActionType.CHANGING` and `ActionType.CHANGED`. In this case the
`data_indices` correspond to the indices of the original data.
2. In case `layer.data` is not populated and new data is not empty use
`ActionType.ADDING` and `ActionType.ADDED`. The `data_indices`
correspond to the data being added.
3. In case new data is empty use `ActionType.REMOVING` and
`ActionType.REMOVED`. The `data_indices` correspond to the original data

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>