Add tests to cover slicing behavior when changing layers and data (#4819)
# References and relevant issues

Related to #6198 and #6199

# Description

This adds tests to cover cases when either layers are added, removed, or
their data changes in ways that should trigger data to be sliced and the
canvas view to be updated. Some of these cases pass due to recent
changes (e.g. #5522), but some fail due to remaining issues (e.g. #6198
and #6199).

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>