Use class name for object that does not have qt name (#6222)
# Description

I have found that when the object does not have a set name, our
performance tracing only provides the event name without information
from which class it comes. So it. is hard to locate the real problem.

This PR add using a class name if the object does not have a name.