Fix typing checks in PRs (#6230)
#6203 was merged without updating the latest changes from `main` and now
the typing checks in PRs fails due to updates in `napari.components`. We
don't run those changes on `main`, though maybe we should given that we
don't requires PR branches to be up-to-date before merging (only that
git reports no conflicts).

This PR fixes the typing failure by casting the type away from the
possible value of `None`, which is a side-effect of calling
`_open_preferences_dialog`. An alternative fix is to have
`_open_preferences_dialog` to return `PreferencesDialog`, which I like
but felt like a bigger change than needed. Of course, we can also use `#
type: ignore`, but I don't there's a need here.