fix outdated dimension sliders documentation (#241)
# References and relevant issues
Closes #240

# Description
This PR updates the documentation under the section `Dimension Sliders`.
Here it was still shown that `current step` would return
a single integer and could be set by assigning a single integer. I have
checked whether this was also the case elsewhere in the documentation,
but that is not the case.

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Peter Sobolewski <pete.sd@gmail.com>