Update color texture build to reduce collisions, and fix collision handling (#6182)
# Description

During work on #6179 I found that the mechanism for avoiding collision
does not work correctly (see tests). I also found that the shader code
do not have a stop condition.

In this PR, I also add a dynamic calculation of the hash table size. I
also decide to use prime numbers as the size of the table to reduce
collision probability.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>