Fix benchmark suite virtualenvs, and disable broken memory benchmarks (#6278)
# Description
Use `[virtualenv]` extras to fix the benchmark setup. 

https://github.com/airspeed-velocity/asv/issues/1323
https://github.com/airspeed-velocity/asv/pull/1324

disable `mem_layer` benchmarks as they do not work on the current main.
It needs to be reenabled in follow-up PR. To have it working, it is
required to have `__sizeof__` fixed, which is part of this commit.

Add benchmark "testing" to pull request test. To spot directly which PR
breaks benchmarks execution.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>