Public API for playback of dimension
This commit adds the functions `play_dim` and `stop_dim` to `viewer.window` to provide a public API for starting and stopping the playback of a dimension. Before, the only API access to the playback functionality was through the `qt_viewer` module (`viewer.window.qt_viewer.dims.play()`). Public access to `qt_viewer` is deprecated and will be removed in `0.5.0`.