Add pystack to debug hanging linux test (#6310)
# Description

[Pystack](https://github.com/bloomberg/pystack) is bloomberg stack
analysis tool for live debugging python process that includes also C
stack.

It has ability to attach to running process and print stactrace. 

The pytest-pystack plugin do this for test. If test is not finished
under given period (60 seconds here) it will print trace to show where
process is.

As we have problem oh hanging test I think it may be worth to try this. 

Unfortunately this solution is linux only.