Maint: mypy opt-in to opt-out.
A number of "stricter" mypy option were opt-in, I swapped the logic and
explicitly opted-out for a few of the relevant errors in the corresponding
files.

The files are grouped together based on the flags that I had to enable.
This is relevant as or example enabling a specific ignore in a file may
effect other ones (typically enabling a flag may have false positive for
warn_unused_ignores).

This should make introducing new files with error more difficult,
as well as picking one file from those list, removing it, running mypy
and fixing the corresponding errors. And add a few other checks.

This include a few fixes as well.