Exclude the loaded property when linking two layers (#6377)
# References and relevant issues
Closes #6372

# Description
This fixes a bug when linking layers and using the experimental async
loading/slicing feature. The bug exists for both the old implementation
available in v0.4.* and the new implementation available on the main
branch, which are the only things that can cause `Layer.loaded` to
change value and therefore cause the corresponding `Layer.events.loaded`
to be emitted. The underlying cause is that the `loaded` property (which
returns `True` when a layer's slice/view is fully loaded) is read-only
from the public API (i.e. there is no public property setter).

The fix is to add `loaded` to the attribute/property names excluded when
finding the common evented attributes. Even if this were user-settable
(which it should *not* be), it should still be excluded because it
corresponds to a slice/view of a layer which is not state that should be
shared between two linked layers.

I added a test to cover this specific case.

I also fixed some nearby typos.