Image layers can't have converted data type using contextual menu, only Labels (#252)
# References and relevant issues
Related to https://github.com/napari/napari/issues/6349
Note this doesn't close that, just makes it a true feature request.

# Description

Currently the viewer tutorial says the contextual menu can be used to
convert Image data type.
It cannot and not just because of the contextual menu setup, but because
the conversions are not implemented for image layers.
In this PR I fix this to say only Labels layers can be converted.