Bugfix: Account for multiscale for labels in 3d (#6317)
Closes https://github.com/napari/napari/issues/6068
In 3D viewer mode, only the lowest resolution of a multiscale image is
supported.
However, this wasn't being accounted for in the `get_value` method when
mousing over a 3D labels layer.
(Note: the Image layer simply doesn't do `get_value` in 3D at all,
regardless of multiscale.)

In this PR, thanks to Lorenzo, I make the following changes:
1. the `data_level` is properly set to the lowest resolution level when
in 3D. When it's not multiscale, the data_level will remain `0`.
2. the bounding box is calculated using the `data_level` rather the full
data extent
3. the sample points for the ray are recomputed accounting for the
downsampling. When it's not multiscale or it's 3D, the downsample factor
is identity.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>