Fix setting shortcuts with modifier keys and only one modifier key as a shortcut on Windows/Linux (#6330)
# References and relevant issues

Closes #6296

# Description

* Fix shortcuts becoming `<Modifier+unknown>` (for example
`Ctrl+unknown`) in the GUI as well as the generation of warnings like
```UserWarning: `Ctrl-+T` does not seem to be a valid shortcut Key.!```
on Windows (and seems like it was happening on Linux too). A preview:


![shortcuts](https://github.com/napari/napari/assets/16781833/d7e6c22c-cb5f-44dd-87b7-f7fbc0649b4b)

* Add more tests for the `ShortcutEditor` widget

* Fix logic to prevent setting more than one modifier key as a shortcut
(for example `Ctrl + Shift`) while allowing to set single modifier keys
as shortcuts (for example `Shift`). To be able to set such a shortcut
you will need to press the modifier key and then press one of the keys
that is not allowed to be set a shortcut [(`Return`, `Tab`, `CapsLock`
or
`Enter`)](https://github.com/napari/napari/blob/7dfd87fbb2fd96cb9fc9f38f28ee5b5a9b85dd42/napari/_qt/widgets/qt_keyboard_settings.py#L653-L662)
A preview:



![shortcuts_only_mod_val](https://github.com/napari/napari/assets/16781833/ef6bb914-16d6-428e-b42f-063141f48dc8)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>