Fix cycle in _update_draw/_set_highlight for Points and Shapes (high CPU background usage) (#6425)
I agree something like https://github.com/napari/napari/issues/6234 is a
better long-term fix, but we should do something in the meantime because
the high CPU is pretty severe.

# References and relevant issues
Closes #6223

# Description
This breaks a cycle when updating the highlight on Points and Shapes
layers. The problem was `_set_highlight` being called with `force=True`
in `_update_draw`. This caused a cycle as redrawing the highlights would
again call `_update_draw`. This was added to fix the appearance of the
highlights during interaction - see #6223 for more discussion. I'm
hoping @brisvag can confirm this PR maintains the correct behavior.

This PR tracks `scale_factor` the same way `selected_data` and other
properties are stored in order to conditionally update the highlight
instead of forcing it.