Fix bounding box transforms when multiscale layer corner goes below zero (#6438)
# References and relevant issues

closes #6394

# Description

There are two differences in handling of multiscale and single scale
one:

Non multiscale:

https://github.com/napari/napari/blob/9d0c449553eaf1acc3be1bf9bc0c8b3eec05afc6/napari/layers/image/_slice.py#L199

Multiscale:

https://github.com/napari/napari/blob/9d0c449553eaf1acc3be1bf9bc0c8b3eec05afc6/napari/layers/image/_slice.py#L236-L241

---------------

Non multiscale 


https://github.com/napari/napari/blob/9d0c449553eaf1acc3be1bf9bc0c8b3eec05afc6/napari/layers/image/_slice.py#L194

Multiscale:


https://github.com/napari/napari/blob/9d0c449553eaf1acc3be1bf9bc0c8b3eec05afc6/napari/layers/image/_slice.py#L244

## So what is the real difference? 

In multiscale data, we only load the visible part of the data. So when
moving the image below 0 coordinates, we do not fetch all data, only
visible places. But this requires updating vispy node transform to place
the image in the correct place still. But then overlays are incorrectly
placed.

So in this PR, I have updated overlays matrices to correct this shift.