Reraise warnings in proxy (#6408)
# Description
While working on debug #6397 I found that when the plugin accesses
qt_viewer the warning message point inside PublicOnlyProxy not in the
plugin code.

It is 
```
/home/czaki/Projekty/napari/napari/utils/_proxies.py:104: FutureWarning: Public access to Window.qt_viewer is deprecated and will be removed in
v0.6.0. It is considered an "implementation detail" of the napari
application, not part of the napari viewer model. If your use case
requires access to qt_viewer, please open an issue to discuss.!
```

instead of 

```
/home/czaki/.pyenv/versions/napari_3.11/lib/python3.11/site-packages/brainrender_napari/napari_atlas_representation.py:24: FutureWarning: Public access to Window.qt_viewer is deprecated and will be removed in
v0.6.0. It is considered an "implementation detail" of the napari
application, not part of the napari viewer model. If your use case
requires access to qt_viewer, please open an issue to discuss.!
```

As we support both interacting with Viewer directly or from a plugin, we
cannot use a single stacklevel. So I implement reraising ow warnings.

It depends on #6396