Fix problem with transform box of multiscale image (#6390)
fix #6382

The problem described in #6392 comes from using `extent_data` for level
0, but the Transform box visual is children of the image visual. When we
change the level of data then we update the transform matrix of the
image visual which leads to increase size of Transform Box.

In this PR I add an extent property that reflects the current data
level. So We get a smaller extent, which is fixed by scaling on the
image layer.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>