Improve flow of install page (#274)
# Description
Update installation guide:
- If I'm installing the bundled app, I don't care about pre-requisites
for installing the Python package, so move that section under the
"Installing as a Pytho package" section
- Modify the header levels, so the second level now only has "Install as
Python package", "Install as bundled app", "Next steps". In particular,
two sections which are only relevant to the Python package route have
been moved down a level to simplify the top level contents.

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>