Bugfix: ensure gamma and opacity are floats (#6498)
# References and relevant issues

At present, despite docstrings stating that gamma and opacity are floats
are not enforced by napari. Worse yet, we set the default in a number of
places to `1` and not `1.0`.
Downstream, vispy does coerce gamma to float so it's not so obvious, but
it creeps up if you interrogate the layer state using either
`viewer.layers[0].gamma` or `viewer.layers[0]._get_state()` which will
return `1` and a type of `int`
See: https://github.com/napari/napari-animation/issues/149
Closes: https://github.com/napari/napari-animation/issues/149
Same can happen to opacity, but it's more subtle because of the way the
slider value is used.
Closes: https://github.com/napari/napari-animation/issues/152

# Description

In this PR I set gamma and opacity defaults to `1.0` where they had been
previously set to `1` and in the `gamma.setter` and `opacity.setter` I
coerce to float so that the user doesn't shoot themselves in the foot
accidentally.